fix: reorganized project structure
