+++ cmd/dourssreader/go.mod
@@ -1,0 +1,5 @@
+module github.com/kdimonych/go_douuarss/cmd/dourssreader
+
+go 1.24.3
+
+require github.com/kdimonych/go_douuarss/lib v1.0.5
