--- cmd/go.mod
@@ -1,5 +1,0 @@
-module go_douuarss
-
-go 1.24.3
-
-require github.com/kdimonych/go_douuarss v1.0.4
