--- go.mod
@@ -1,3 +1,0 @@
-module github.com/kdimonych/go_douuarss
-
-go 1.24.3
