+++ go.work
@@ -1,0 +1,8 @@
+go 1.24.4
+
+use (
+	./lib
+	./services/gouse_migrate
+	./services/news_checker
+	./cmd/dourssreader
+)
