+++ lib/go.mod
@@ -1,0 +1,3 @@
+module github.com/kdimonych/go_douuarss/lib
+
+go 1.24.3
