+++ lib/rss/model.go
