+++ services/gouse_migrate/go.mod
@@ -1,4 +1,4 @@
-module github.com/kdimonych/go_douuarss/gouse_migrate
+module github.com/kdimonych/go_douuarss/service/gouse_migrate
 
 go 1.24.3
