+++ services/news_checker/go.mod
@@ -1,5 +1,5 @@
-module github.com/kdimonych/go_douuarss/gouse_migrate
+module github.com/kdimonych/go_douuarss/service/news_checker
 
 go 1.24.3
 
-require github.com/kdimonych/go_douuarss v1.0.4
+require github.com/kdimonych/go_douuarss/lib v1.0.4
